Fix sidebar breaking if you don't hit close
If you hit escape on desktop (or back on mobile? someone talked about
this in the discord) then the sidebar would stop working.

This doesn't feel like the best fix, but it does work and doesn't seem
to cause problems. I'm not sure what to do with all the buttons that
close the box on click.