When viewing fish as a table, filter each set independently
Because fish have multiple location, weather, season entries, there are
often entries that make no sense in the table for some fish.

For example, tilapia will show the 'Cave Mining Earth' row even if
location is filtering to 'Lake Temple'. This change addresses this
issue by moving the splitting of fish up to the level where we can
then apply the filter to each location set.

Now instad of matching any of the fish location sets and then showing
all of them, it match each set independently and shows that one only if
it matches.